+++ R/eovcheck.R
@@ -305,20 +305,6 @@ eovcheckBase = function(diagnosticInfo, xlab, ylab, col, smoother, twosd, ...) {
 eovcheckGgplot2 = function(diagnosticInfo, xlab, ylab, col, smoother, twosd) {
   requirePlottingPackage("ggplot2")
 
-  ggplotFunctions = getPlottingFunctions(
-    "ggplot2",
-    c(
-      "ggplot", "aes", "geom_point", "geom_hline",
-      "geom_smooth", "labs", "annotate"
-    )
-  )
-  ggplot = ggplotFunctions[["ggplot"]]
-  aes = ggplotFunctions[["aes"]]
-  geomPoint = ggplotFunctions[["geom_point"]]
-  geomHline = ggplotFunctions[["geom_hline"]]
-  geomSmooth = ggplotFunctions[["geom_smooth"]]
-  labs = ggplotFunctions[["labs"]]
-  annotate = ggplotFunctions[["annotate"]]
 
   diagnosticData = diagnosticInfo$diagnosticData
   plotData = data.frame(
@@ -330,8 +316,8 @@ eovcheckGgplot2 = function(diagnosticInfo, xlab, ylab, col, smoother, twosd) {
     plotData,
     aes(x = plotData[["fitted"]], y = plotData[["residuals"]])
   ) +
-    geomPoint(shape = 1) +
-    geomHline(yintercept = 0, linetype = 3, colour = "lightgrey") +
+    geom_point(shape = 1) +
+    geom_hline(yintercept = 0, linetype = 3, colour = "lightgrey") +
     labs(x = xlab, y = ylab, title = "")
 
   if (smoother) {
@@ -340,12 +326,12 @@ eovcheckGgplot2 = function(diagnosticInfo, xlab, ylab, col, smoother, twosd) {
     } else {
       col
     }
-    plotObject = plotObject + geomSmooth(method = "loess", se = FALSE, colour = smootherColour)
+    plotObject = plotObject + geom_smooth(method = "loess", se = FALSE, colour = smootherColour)
   }
 
   if (twosd) {
     sigma = summary(diagnosticInfo$fit)$sigma
-    plotObject = plotObject + geomHline(yintercept = c(-2, 2) * sigma, linetype = 3, colour = "grey")
+    plotObject = plotObject + geom_hline(yintercept = c(-2, 2) * sigma, linetype = 3, colour = "grey")
   }
 
   if (diagnosticInfo$showLevene) {