+++ tests/testthat/helper-internal-helpers.R
@@ -1,0 +1,3 @@
+getS20xInternal = function(name) {
+  getFromNamespace(name, "s20x")
+}